Fix: eslint errors
