+++ .eslintrc
@@ -1,0 +1,12 @@
+{
+    "extends": "airbnb",
+    "plugins": [
+        "react",
+        "jsx-a11y",
+        "jsx",
+        "import"
+    ],
+    "rules": {
+        "react/prop-types": 0
+    }
+}
