+++ .eslintrc.json
@@ -1,0 +1,14 @@
+{
+    "extends": "airbnb",
+    "env": {
+      "browser": true
+    },
+    "plugins": [
+      "react",
+      "jsx-a11y",
+      "import"
+    ],
+    "rules": {
+      "react/prop-types": 0
+    }
+}
