+++ src/App.test.jsx
@@ -1,3 +1,5 @@
+/* eslint-disable no-undef */
+
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
