+++ src/index.js
@@ -1,9 +1,12 @@
+/* eslint-disable react/jsx-filename-extension */
+/* eslint-disable no-undef */
+
 import React from 'react';
 import ReactDOM from 'react-dom';
+import 'normalize.css';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
 
-import 'normalize.css'
 import './index.css';
 
 ReactDOM.render(<App />, document.getElementById('root'));
