+++ src/registerServiceWorker.js
@@ -1,3 +1,4 @@
+/* eslint-disable */
 // In production, we register a service worker to serve assets from local cache.
 
 // This lets the app load faster on subsequent visits in production, and gives
